fix: widen deleted_ids from List<Int> to List<Long> across all modules
The server sends article IDs as i64. The new deleted_ids field was
introduced as List<Int> in the Kotlin contract, which would fail at
runtime if an ID exceeded Int.MAX_VALUE. Widen to List<Long> in shared
models, ArticleStore interface, and all implementations (Room, IndexedDB)
and their tests. IndexedDB's delete() call now converts Long to Double
since Kotlin/JS Long is a wrapper object, not a native JS number.

+++ app/src/main/java/eu/monniot/feed/store/ArticleStoreDao.kt
@@ -21,7 +21,7 @@ interface ArticleStoreDao {
     suspend fun upsert(articles: List<SyncArticleEntity>)
 
     @Query("DELETE FROM sync_articles WHERE id IN (:ids)")
-    suspend fun deleteByIds(ids: List<Int>)
+    suspend fun deleteByIds(ids: List<Long>)
 
     @Query("DELETE FROM sync_articles")
     suspend fun clearArticles()
